Moved retry_stale_element_reference_exception decorator to utils.py
