+++ pages/webview/content.py
@@ -4,10 +4,9 @@
 
 import re
 
-from functools import wraps
-
 from selenium.webdriver.common.by import By
-from selenium.common.exceptions import StaleElementReferenceException
+
+from tests.utils import retry_stale_element_reference_exception
 
 from pages.webview.base import Page
 from regions.webview.base import Region
@@ -41,30 +40,9 @@ class Content(Page):
     def share(self):
         return self.content_header.share
 
-    # Retries StaleElementReferenceExceptions up to n-1 times (default n=3)
-    # Decorator with optional argument based on: https://stackoverflow.com/a/3931903
-    def retry_stale_element_reference_exception(method_or_max_tries):
-        def wrap(method):
-            @wraps(method)
-            def wrapper(*args, **kwargs):
-                for i in range(max_tries):
-                    try:
-                        return method(*args, **kwargs)
-                    except StaleElementReferenceException:
-                        if i >= max_tries - 1:
-                            raise
-            return wrapper
-
-        if callable(method_or_max_tries):
-            max_tries = 3
-            return wrap(method_or_max_tries)
-        else:
-            max_tries = method_or_max_tries
-            return wrap
-
     # The media-header div can be reloaded at seemingly random times
     # Any method that accesses an element inside this header
-    # must retry StaleElementReferenceExceptions using the above decorator
+    # must retry StaleElementReferenceExceptions using retry_stale_element_reference_exception
     @property
     @retry_stale_element_reference_exception
     def is_section_title_displayed(self):
